plugins/guardian-onboarding/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=./plugins/guardian-onboarding/requirements.txt ./plugins/guardian-onboarding/requirements.in
#
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
filelock==3.24.3
    # via huggingface-hub
fsspec==2026.2.0
    # via huggingface-hub
hf-xet==1.2.0
    # via huggingface-hub
huggingface-hub==0.36.2
    # via -r plugins/guardian-onboarding/requirements.in
idna==3.11
    # via requests
packaging==26.0
    # via huggingface-hub
python-dotenv==1.0.1
    # via -r plugins/guardian-onboarding/requirements.in
pyyaml==6.0.3
    # via huggingface-hub
requests==2.32.5
    # via huggingface-hub
tqdm==4.67.3
    # via huggingface-hub
typing-extensions==4.15.0
    # via huggingface-hub
urllib3==2.6.3
    # via requests
